requirements.txt
gunicorn
flask
flask_sqlalchemy
Flask-RESTful
SQLAlchemy
flask_script
flask_migrate
psycopg2
flask_cors
flask_moment
Werkzeug
babel
python-jose-cryptodome